refactor(update): inline version formatting
