+++ src/lib/i18n.tsx
@@ -102,4 +102,15 @@ export function useI18n(): LocaleContextType {
   return context;
 }
 
-export { formatMessage } from "./message-format";
+// Helper function to replace placeholders in strings
+// Example: formatMessage("Hello {name}", { name: "World" }) => "Hello World"
+export function formatMessage(
+  template: string,
+  values?: Record<string, string | number>
+): string {
+  if (!values) return template;
+  return template.replace(
+    /\{(\w+)\}/g,
+    (_, key) => String(values[key] ?? `{${key}}`)
+  );
+}
