+++ src/pages/Settings.tsx
@@ -7,8 +7,7 @@ import ImportHistoryModal from "../components/ImportHistoryModal";
 import { ensureGatewayInitialized, refreshEngineConfigState } from "../lib/engine-bootstrap";
 import { ChannelManagementSettings } from "../components/ChannelManagementSettings";
 import { TerminalSettingsSection } from "../components/TerminalSettingsSection";
-import { useI18n } from "../lib/i18n";
-import { formatUpdateAvailableMessage, hasUpdateVersion } from "../lib/update-message";
+import { formatMessage, useI18n } from "../lib/i18n";
 import { logger } from "../lib/logger";
 import { useAuthGuard } from "../lib/useAuthGuard";
 import { isElectron } from "../lib/platform";
@@ -229,8 +228,9 @@ export default function Settings() {
       return;
     }
     if (result.status === "available" || result.status === "downloading" || result.status === "downloaded") {
-      if (hasUpdateVersion(result.version)) {
-        setAvailableUpdateVersion(result.version.trim());
+      const version = result.version?.trim();
+      if (version) {
+        setAvailableUpdateVersion(version);
         setUpdateCheckStatus("available");
       } else {
         setUpdateCheckStatus("error");
@@ -995,10 +995,7 @@ export default function Settings() {
                       </Show>
                       <Show when={updateCheckStatus() === "available"}>
                         <p class="text-xs text-blue-600 dark:text-blue-400 mt-1">
-                          {formatUpdateAvailableMessage(
-                            t().update.available,
-                            availableUpdateVersion(),
-                          )}
+                          {formatMessage(t().update.available, { version: availableUpdateVersion() })}
                         </p>
                       </Show>
                       <Show when={updateCheckStatus() === "error"}>
